creates app folder and moves components, modules, state folders into app creates styles.js and styles folder to contain style sheet refactors styles out of components/into style sheet
