+++ index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Provider } from 'react-redux'
-import store from './state/store/store'
+import store from './app/state/store/store'
 import App from './App';
 import { registerRootComponent } from 'expo';
